EDM-119: Allow decimals for  hours_worked_per_week (#2405)
* Allow partial hours worked per week

* Update openapi.yaml

* Fix lint, build

* Revert make format sort changes

* Update tests

* Update tests

* Fix lint

+++ api/massgov/pfml/api/models/claims/common.py
@@ -1,4 +1,5 @@
 from datetime import date
+from decimal import Decimal
 from typing import List, Optional
 
 from massgov.pfml.api.models.common import PreviousLeave
@@ -49,7 +50,7 @@ class EmployerClaimReview(PydanticBaseModel):
 
     comment: Optional[str]
     employer_benefits: List[EmployerBenefit]
-    hours_worked_per_week: Optional[int]
+    hours_worked_per_week: Optional[Decimal]
     previous_leaves: List[PreviousLeave]
     employer_decision: Optional[str]
     fraud: Optional[str]
